+++ protected/humhub/modules/user/models/User.php
@@ -16,6 +16,7 @@ use humhub\modules\user\events\UserEvent;
 use humhub\modules\friendship\models\Friendship;
 use humhub\modules\space\models\Space;
 use humhub\modules\content\models\Content;
+use humhub\modules\space\models\Membership;
 
 /**
  * This is the model class for table "user".
@@ -338,11 +339,10 @@ class User extends ContentContainerActiveRecord implements \yii\web\IdentityInte
      */
     public function beforeDelete()
     {
-
         // We don't allow deletion of users who owns a space - validate that
-        foreach (\humhub\modules\space\models\Membership::GetUserSpaces($this->id) as $space) {
+        foreach (Membership::GetUserSpaces($this->id, false) as $space) {
             if ($space->isSpaceOwner($this->id)) {
-                throw new Exception('Tried to delete a user (' . $this->id . ') which is owner of a space!');
+                throw new Exception('Tried to delete a user (' . $this->id . ') which is owner of a space (' . $space->id . ')!');
             }
         }