+++ protected/humhub/docs/CHANGELOG.md
@@ -16,6 +16,10 @@ migrate your code to SpreadsheetExport. PHPOffice is replaced by PHPSpreadsheet.
 - Fix: Typo in password recovery (@acs-ferreira)
 - Fix: Profile posts of friends not appears on dashboard
 - Fix #2745: Yii2 2.0.13 will break the admin interface
+- Enh: Allow auto detection of response dataType
+- Fix #2947: Allow json success result on modalSubmit
+- Enh: Disabled automatic content following on likes by default
+- Enh: Improved IntegrityChecker memory usage
 - Chg: `PendingRegistrationsController->createCVS` removed
 
 1.2.4  (December 13, 2017)